Делит модуль util.js на два модуля
- Вместо одного модуля с утилитарными функциями, в проекте создана папка
  для разделения кода из util.js и будущих вспомогательных функций.
- Удален модуль util.js, код из него разнесен в модули render.js и
  common.js
- render.js отвечает за функцию отрисовки элемента, создания элемента,
  замены элементов и удаления элементов. Сама функция render изменена,
  теперь вместо того чтобы передавать ей DOM элемент через вызов на
  экземпляре класса метода getElement(), функция может принимать
  компоненты как в качестве контейнера для отрисовки, так и в качестве
  самого объекта отрисовки.
- common.js отвечает за вспомогательные функции, в виде получения
  случайного числа и получения продолжитильности между двумя отрезками
  времени.
