+++ src/view/abstract.js
@@ -15,7 +15,7 @@ export default class Abstract {
 
   getElement () {
     if (!this._element) {
-      this._element = createElement(this._element.getTemplate());
+      this._element = createElement(this.getTemplate());
     }
 
     return this._element;
